- Fixed tokenizer bug where floats with uppercase E and F would not be recognized as float values - Changed GeometricApproximation() to accept any token value - Added 2 new test files (one renders with the current version, the other needs "Polygon" to be implemented) - Improved error report.. now tells at what line of the scene file there is a problem

+++ RI_System/src/RI_Primitive_Poly.cpp
@@ -35,7 +35,7 @@ Polygon::Polygon( ParamList &params, const Vector3 *pVertsP, int vertsN ) :
 {
 	mVertsP.resize( vertsN );
 	
-	for (size_t i=0; i < vertsN; ++i)
+	for (size_t i=0; i < (size_t)vertsN; ++i)
 		mVertsP[i] = pVertsP[i];
 }
 
@@ -43,6 +43,8 @@ Polygon::Polygon( ParamList &params, const Vector3 *pVertsP, int vertsN ) :
 void Polygon::Render( GState &gstate )
 {
 	PUTPRIMNAME( "* Polygon" );
+	
+	//mVertsP
 
 }
